+++ src/index.ts
@@ -6,6 +6,8 @@ export { default as useFormContext } from './useFormContext';
 
 export { default as useField } from './useField';
 
+export { default as useFieldValue } from './useFieldValue';
+
 export { default as useCheckboxField } from './useCheckboxField';
 
 export { get } from './utils';