+++ src/useFieldValue.ts
@@ -1,0 +1,5 @@
+import useField from './useField';
+
+const useFieldValue = (name: string) => useField(name).field.value;
+
+export default useFieldValue;